Reorganized a bit.
