+++ metron-contrib/metron-performance/src/main/java/org/apache/metron/performance/load/monitor/EPSGeneratedMonitor.java
@@ -15,36 +15,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.metron.performance.load;
+package org.apache.metron.performance.load.monitor;
 
 import java.util.Optional;
-import java.util.function.Supplier;
+import java.util.concurrent.atomic.AtomicLong;
 
-public abstract class AbstractMonitor implements Supplier<Long> {
-  private static final double EPSILON = 1e-6;
-  protected Optional<?> kafkaTopic;
-  protected long timestampPrevious = 0;
-  public AbstractMonitor(Optional<?> kafkaTopic) {
-    this.kafkaTopic = kafkaTopic;
+public class EPSGeneratedMonitor extends AbstractMonitor {
+  private AtomicLong numSent;
+  private long numSentPrevious = 0;
+  public EPSGeneratedMonitor(Optional<?> kafkaTopic, AtomicLong numSent) {
+    super(kafkaTopic);
+    this.numSent = numSent;
   }
-
-  protected abstract Long monitor(double deltaTs);
 
   @Override
-  public Long get() {
-    long timeStarted = System.currentTimeMillis();
-    Long ret = null;
-    if(timestampPrevious > 0) {
-      double deltaTs = (timeStarted - timestampPrevious) / 1000.0;
-      if (Math.abs(deltaTs) > EPSILON) {
-        ret = monitor(deltaTs);
-      }
+  protected Long monitor(double deltaTs) {
+    if(kafkaTopic.isPresent()) {
+      long totalProcessed = numSent.get();
+      long written = (totalProcessed - numSentPrevious);
+      long epsWritten = (long) (written / deltaTs);
+      numSentPrevious = totalProcessed;
+      return epsWritten;
     }
-    timestampPrevious = timeStarted;
-    return ret;
+    return null;
   }
 
-  public abstract String format();
+  @Override
+  public String format() {
+    return "%d eps generated to " + kafkaTopic.get();
+  }
 
-  public abstract String name();
 }
